Minor Page Changes
ProcessPage functions must return a string (to facilitate error handling - the functions have a LOT of different error cases and writing WriteLine a whole bunch of times is a pain).
ProcessPages ask you to confirm your decision to exit - the default is no, since you will lose all your progress.
Menupages display the description - it is a bit longer and more descriptive than the name, and should NOT just be "chose an option," because the menu displays that automatically.
Other than that, it's mostly added documentation. I changed a few variables names because they didn't seem intuitive and I marked a few things for review.
I kept accidentally hitting enter when navigating menus and having to back track - I really think this could be problematic, but I will hand it off to my roommates to see if they have any thoughts.